requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.2
gunicorn==20.0.4
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
